chore(proof/tee): remove unused executor, providers, and config modules from base-enclave
Remove ~3,200 lines of dead code from base-enclave and base-enclave-client.
The executor/ and providers/ modules formed an internal execution pipeline
that was never called by the nitro server (which uses base-proof-client
instead). The config/defaults module was only consumed internally.

Relocated externally-used items before deletion:
- ExecutionWitness -> types/witness.rs
- l2_block_to_block_info() -> types/l2_block_ref.rs (with new L2BlockRefError)

Also removed:
- Dead EnclaveProvider trait from base-enclave-client
- 12 unused Cargo.toml dependencies (revm, base-revm, alloy-evm, etc.)
- 6 unused alloy re-exports from lib.rs
- Unused error types (ConfigError, ExecutorError, EnclaveError)
