+++ crates/proof/tee/core/src/types/mod.rs
@@ -2,12 +2,16 @@ mod account;
 pub use account::{AccountResult, StorageProof};
 
 mod config;
-pub use config::{
-    BlockId, Genesis, GenesisSystemConfig, MARSHAL_BINARY_SIZE, PerChainConfig, RollupConfig,
-};
+pub use config::{BlockId, Genesis, GenesisSystemConfig, PerChainConfig, RollupConfig};
+
+mod l2_block_ref;
+pub use l2_block_ref::{L2BlockRefError, l2_block_to_block_info};
 
 mod output;
 pub use output::{output_root_v0, output_root_v0_with_hash};
 
 mod rpc;
 pub use rpc::{AggregateRequest, ExecuteStatelessRequest};
+
+mod witness;
+pub use witness::ExecutionWitness;
